*add noise terrain generation

+++ game/systems/utils/FPSCounterSystem.cpp
@@ -38,7 +38,7 @@ void CFPSCounterSystem::OnUpdate(registry_t & Registry_, float Delta_)
 
   if (m_DeltaAccumulator >= FPS_COUNTER_UPDATE_INTERVAL)
   {
-    spdlog::info("FPS: {}, Entities: {}", static_cast<float>(m_FramesCount), Registry_.size());
+    spdlog::info("FPS: {}, Entities: {}", m_FramesCount, Registry_.size());
 
     m_DeltaAccumulator = 0;
     m_FramesCount      = 0;
